Split paginator.rb file
