+++ lib/pliny.rb
@@ -6,6 +6,8 @@ require_relative "pliny/errors"
 require_relative "pliny/extensions/instruments"
 require_relative "pliny/helpers/encode"
 require_relative "pliny/helpers/paginator"
+require_relative "pliny/helpers/paginator/paginator"
+require_relative "pliny/helpers/paginator/integer_paginator"
 require_relative "pliny/helpers/params"
 require_relative "pliny/log"
 require_relative "pliny/request_store"
